reusable pipelines refactored into a separate module
